fix types conflict. Closes #1

+++ internal/domain/model/transaction.go
@@ -1,7 +1,7 @@
 package model
 
 type Transaction struct {
-	From   string `json:"from"`
-	To     string `json:"to"`
-	Amount int    `json:"amount"`
+	From   string  `json:"from"`
+	To     string  `json:"to"`
+	Amount float64 `json:"amount"`
 }
